refactor: add src for consumers, nodes, and providers
