+++ beine_cpp/src/node/legs_node.cpp
@@ -18,54 +18,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef BEINE_CPP__PROVIDER__STANCE_PROVIDER_HPP_
-#define BEINE_CPP__PROVIDER__STANCE_PROVIDER_HPP_
+#include <beine_cpp/node/legs_node.hpp>
 
-#include <rclcpp/rclcpp.hpp>
-
-#include "../node.hpp"
-#include "../utility.hpp"
+#include <string>
 
 namespace beine_cpp
 {
 
-class StanceProvider : public LegsNode
+LegsNode::Options::Options()
+: legs_prefix("/legs")
 {
-public:
-  struct Options : public virtual LegsNode::Options
-  {
-  };
+}
 
-  inline explicit StanceProvider(rclcpp::Node::SharedPtr node, const Options & options = Options());
-
-  void set_stance(const Stance & stance);
-
-  const Stance & get_stance() const;
-
-private:
-  rclcpp::Publisher<StanceMsg>::SharedPtr stance_publisher;
-
-  Stance current_stance;
-};
-
-StanceProvider::StanceProvider(
-  rclcpp::Node::SharedPtr node, const StanceProvider::Options & options)
-: LegsNode(node, options)
+LegsNode::LegsNode(rclcpp::Node::SharedPtr node, const LegsNode::Options & options)
+: node(node),
+  legs_prefix(options.legs_prefix)
 {
-  // Initialize the stance publisher
-  {
-    stance_publisher = get_node()->create_publisher<StanceMsg>(
-      get_legs_prefix() + STANCE_SUFFIX, 10);
+}
 
-    RCLCPP_INFO_STREAM(
-      get_node()->get_logger(),
-      "Stance publisher initialized on " << stance_publisher->get_topic_name() << "!");
-  }
+rclcpp::Node::SharedPtr LegsNode::get_node() const
+{
+  return node;
+}
 
-  // Initial data publish
-  set_stance(get_stance());
+const std::string & LegsNode::get_legs_prefix() const
+{
+  return legs_prefix;
 }
 
 }  // namespace beine_cpp
-
-#endif  // BEINE_CPP__PROVIDER__STANCE_PROVIDER_HPP_
