+++ beine_cpp/src/utility/stance.cpp
